feat: remove v1 agent create and generate spec commands
