+++ schemas/agent-create.json
@@ -4,18 +4,142 @@
   "definitions": {
     "AgentCreateResult": {
       "type": "object",
+      "additionalProperties": false,
       "properties": {
+        "previewFilePath": {
+          "type": "string"
+        },
         "isSuccess": {
           "type": "boolean"
         },
         "errorMessage": {
           "type": "string"
+        },
+        "agentId": {
+          "type": "object",
+          "properties": {
+            "botId": {
+              "type": "string"
+            },
+            "botVersionId": {
+              "type": "string"
+            },
+            "plannerId": {
+              "type": "string"
+            }
+          },
+          "required": ["botId", "botVersionId", "plannerId"],
+          "additionalProperties": false,
+          "description": "If the agent was created with saveAgent=true, these are the IDs that make up an agent; Bot, BotVersion, and GenAiPlanner metadata."
+        },
+        "agentDefinition": {
+          "type": "object",
+          "properties": {
+            "agentDescription": {
+              "type": "string"
+            },
+            "topics": {
+              "type": "array",
+              "items": {
+                "type": "object",
+                "properties": {
+                  "scope": {
+                    "type": "string"
+                  },
+                  "topic": {
+                    "type": "string"
+                  },
+                  "actions": {
+                    "type": "array",
+                    "items": {
+                      "type": "object",
+                      "properties": {
+                        "actionName": {
+                          "type": "string"
+                        },
+                        "exampleOutput": {
+                          "type": "string"
+                        },
+                        "actionDescription": {
+                          "type": "string"
+                        },
+                        "inputs": {
+                          "type": "array",
+                          "items": {
+                            "type": "object",
+                            "properties": {
+                              "inputName": {
+                                "type": "string"
+                              },
+                              "inputDataType": {
+                                "type": "string"
+                              },
+                              "inputDescription": {
+                                "type": "string"
+                              }
+                            },
+                            "required": ["inputName", "inputDataType", "inputDescription"],
+                            "additionalProperties": false
+                          },
+                          "minItems": 1,
+                          "maxItems": 1
+                        },
+                        "outputs": {
+                          "type": "array",
+                          "items": {
+                            "type": "object",
+                            "properties": {
+                              "outputName": {
+                                "type": "string"
+                              },
+                              "outputDataType": {
+                                "type": "string"
+                              },
+                              "outputDescription": {
+                                "type": "string"
+                              }
+                            },
+                            "required": ["outputName", "outputDataType", "outputDescription"],
+                            "additionalProperties": false
+                          },
+                          "minItems": 1,
+                          "maxItems": 1
+                        }
+                      },
+                      "required": ["actionName", "exampleOutput", "actionDescription", "inputs", "outputs"],
+                      "additionalProperties": false
+                    },
+                    "minItems": 1,
+                    "maxItems": 1
+                  },
+                  "instructions": {
+                    "type": "array",
+                    "items": {
+                      "type": "string"
+                    }
+                  },
+                  "classificationDescription": {
+                    "type": "string"
+                  }
+                },
+                "required": ["scope", "topic", "actions", "instructions", "classificationDescription"],
+                "additionalProperties": false
+              },
+              "minItems": 1,
+              "maxItems": 1
+            },
+            "sampleUtterances": {
+              "type": "array",
+              "items": {
+                "type": "string"
+              }
+            }
+          },
+          "required": ["agentDescription", "topics", "sampleUtterances"],
+          "additionalProperties": false
         }
       },
-      "required": [
-        "isSuccess"
-      ],
-      "additionalProperties": false
+      "required": ["agentDefinition", "isSuccess"]
     }
   }
 }
